refactor(docs-gen): modularize extension extractor
