+++ infra/docs-gen/package.json
@@ -15,8 +15,8 @@
   },
   "scripts": {
     "build": "node src/generate-docs.mjs && yfm -i ../../tmp/docs-src -o ../../dist/docs",
-    "extract:names": "node src/extract-extension-names.mjs",
-    "test": "node --test src/*.test.mjs"
+    "extract": "node src/extract-extension-data.mjs",
+    "test": "node --test src/extractor/*.test.mjs"
   },
   "dependencies": {
     "@diplodoc/cli": "5.43.0",
